+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.5.0] - 2023-07-20
 ### Added
 - `Emitter::new()` and `Emitter::into_error()` to enable using the Emitter manually.
 - Implemented `Extend` for `Emitter` and `Error`.
@@ -40,7 +41,8 @@ Only documentation changes.
 ## [v0.1.0] 
 **Initial Release**
 
-[unreleased]: https://github.com/ModProg/manyhow/compare/v0.4.2...HEAD
+[unreleased]: https://github.com/ModProg/manyhow/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/ModProg/manyhow/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/ModProg/manyhow/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/ModProg/manyhow/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/ModProg/manyhow/compare/v0.3.0...v0.4.0